requirements.txt
flask
flask-sqlalchemy
yfinance
pytesseract
pandas
openpyxl
pillow
easyocr
python-dotenv
